+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Fix taginput when ``allow-new`` is enabled
 * Fix pagination order when simple
 * #1126 Fix init sort using columns table prop
+* #898, #1154 Fix browser autocomplete using multiple fields
 
 ## 0.7.1